% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\docType{data}
\name{DBT}
\alias{DBT}
\title{A decoding byte table to map raw intergers to genotype dosages.}
\description{
The table avoids bit shifting and may speed up the reading of plink BED.
}
\details{
The decoding table approach is still experimental.

Actually, DBT == dbd(as.raw(seq(0x00, 0xFF)), 4L)
}
