% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{gid}
\alias{gid}
\title{Infer Sample ID from a symmetric matrix}
\usage{
gid(x, sep = ".")
}
\arguments{
\item{x}{matrix}

\item{sep}{separator between FID and IID forming the sample ID}
}
\value{
data.frame of inferred family ID and individual ID.
}
\description{
Exam the row name for family and individual id.
}
\details{
For matrices without rowname, id are automatically generated.

By common practice, the  row names or a matrix are in  the form of [FID.]IID.
Samples without family ID are given one identical to their individual ID.
}
