% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp.R
\name{lc}
\alias{lc}
\title{Line Count}
\usage{
lc(f)
}
\arguments{
\item{f}{the file name, or a connection.}
}
\description{
Count the occurance of '\\n', much faster than \code{readLine} and
\code{read.table}. Although slower than the unix command "wc -l",
it upholds platform independency.
}
