% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{readFAM}
\alias{readFAM}
\title{Read FAM file}
\usage{
readFAM(pfx)
}
\arguments{
\item{pfx}{prefix of a PLINK file set.}
}
\value{
data frame describing individuals,  loaded from the FAM file.
}
\description{
Read FAM file
}
\examples{
pfx <- file.path(system.file("extdata", package="plinkFile"), "m20")
bed <- readBED(pfx, row=1, col=1, quiet=FALSE)
bed

}
