% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{saveBED}
\alias{saveBED}
\title{Save BED file}
\usage{
saveBED(pfx, bed, quiet = TRUE)
}
\arguments{
\item{pfx}{prefix of the output file set, in PLINK1 BED format.}

\item{bed}{N x P genotype matrix}

\item{quiet}{do not report (def=TRUE)}
}
\description{
Save a R matrix into a PLINK BED file.
}
\details{
This is meant for genotype small enough  to fit into system memory. The size
of R matrix is 16  times the size of the BED file.
}
