% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{testReadBED}
\alias{testReadBED}
\title{Test BED Reader}
\usage{
testReadBED()
}
\description{
Read m20 (bed, bim, and fam) under  "extdata" and compare with the content in
text file "i10.txt" converted from m20 by PLINK.
}
