% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{testReadBSM}
\alias{testReadBSM}
\title{Test Genetic Relatedness Matrix Reader}
\usage{
testReadBSM()
}
\description{
Compare the read from genetic relatedness matrix created from the same genome
segment but stored in different shapes and types.
}
