% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{readBIM}
\alias{readBIM}
\title{Read BIM file}
\usage{
readBIM(pfx)
}
\arguments{
\item{pfx}{prefix of a PLINK file set.}
}
\value{
data frame describing genome variants, loaded from the BIM file.
}
\description{
Read BIM file
}
